Build create and populate function
After some trying I could only make this work by
doing both actions at the same time. Also just
made more sense. When the form is submitted the
object is first created, then is used to create the card
which is linked to its array position with a data
attribute.