Finish styling form.
Got the form styled thematically with the rest of the
page. Really just needed colors and some tweaks to the
inputs. Also was testing the remove function and it
appears to be working? I cannot recreate the issue i had
yesterday...either it is a bug with dev tools or i was
just tired and tripping. Anyways this means im done with
this project ayy.